[Refactor] Try to speed up the compilation through file splitting.
